+++ jpa/src/main/java/de/taimos/dvalin/jpa/config/DatabaseConfig.java
@@ -9,9 +9,9 @@ package de.taimos.dvalin.jpa.config;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -20,12 +20,11 @@ package de.taimos.dvalin.jpa.config;
  * #L%
  */
 
-import javax.persistence.EntityManagerFactory;
-import javax.sql.DataSource;
-
+import liquibase.integration.spring.SpringLiquibase;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.DependsOn;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.jdbc.datasource.init.DataSourceInitializer;
 import org.springframework.jdbc.datasource.init.ResourceDatabasePopulator;
@@ -35,7 +34,8 @@ import org.springframework.orm.jpa.vendor.Database;
 import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 
-import liquibase.integration.spring.SpringLiquibase;
+import javax.persistence.EntityManagerFactory;
+import javax.sql.DataSource;
 
 
 @Configuration
@@ -56,6 +56,7 @@ public class DatabaseConfig {
 
 
     @Bean
+    @DependsOn("liquibase")
     public EntityManagerFactory entityManagerFactory(DataSource dataSource) {
         LocalContainerEntityManagerFactoryBean bean = new LocalContainerEntityManagerFactoryBean();
         bean.setDataSource(dataSource);
@@ -76,7 +77,7 @@ public class DatabaseConfig {
     }
 
 
-    @Bean
+    @Bean("liquibase")
     public SpringLiquibase liquibase(DataSource dataSource) {
         SpringLiquibase liquibase = new SpringLiquibase();
         liquibase.setDataSource(dataSource);